# docs/swagger.yaml
definitions:
  request.APIUpdateAccountDetail:
    properties:
      address:
        description: Address is either an EVM or Cosmos address of the account.
        type: string
      monitored:
        description: Monitored indicates whether the service should monitor this account.
        type: boolean
      name:
        description: Name is the identity of the account.
        type: string
    required:
    - address
    - name
    type: object
  request.APIUpdateTokenDetail:
    properties:
      Decimals:
        description: Decimals is the number of decimal places of the token.
        type: integer
      TokenAddress:
        description: TokenAddress is the EVM address of the token.
        type: string
      TokenName:
        description: TokenName is the name of the token.
        type: string
      WhaleDefinition:
        description: WhaleDefinition is the amount to trigger the whale alerts.
        type: number
    required:
    - Decimals
    - TokenAddress
    - TokenName
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    email: support@stellalab.com
    name: StellaLab
    url: https://www.stellalab.com
  description: This service is dedicated to administration of a Whale Alert service.
  title: Whale Alert Admin APIs
  version: "1.0"
paths:
  /admin/account/all:
    get:
      consumes:
      - application/json
      description: Get all stored account details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all account details
  /admin/account/all-monitored:
    get:
      consumes:
      - application/json
      description: Retrieve all the monitored accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all monitored accounts
  /admin/account/update:
    post:
      consumes:
      - application/json
      description: Store the detail of an account
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.APIUpdateAccountDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Account Detail
  /admin/token/all:
    get:
      consumes:
      - application/json
      description: Get all stored token details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all token details
  /admin/token/update:
    post:
      consumes:
      - application/json
      description: Store the detail of a token
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.APIUpdateTokenDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Token Detail
securityDefinitions:
  ApiKeyAuth:
    description: Use for admin authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"